Update votebox to latest order when new vote is cast MU-179

+++ src/src/main/java/com/hcmus/mentor/backend/service/impl/ReminderServiceImpl.java
@@ -27,23 +27,6 @@ public class ReminderServiceImpl implements ReminderService {
     private final GroupRepository groupRepository;
     private final MailService mailService;
 
-    //    public List<Reminder> getReminders(IRemindableService remindableService){
-    //        return remindableService.findReminderToday();
-    //    }
-
-    //    public List<Reminder> saveReminders(){
-    //        List<Reminder> reminders = new ArrayList<>();
-    //        List<Reminder> meetingReminders = getReminders(meetingService);
-    //        List<Reminder> taskReminders = getReminders(taskService);
-    //
-    //        reminders.addAll(meetingReminders);
-    //        reminders.addAll(taskReminders);
-    //
-    //        reminderRepository.saveAll(reminders);
-    //
-    //        return reminders;
-    //    }
-
     @Override
     public void sendReminders() throws FirebaseMessagingException {
         List<Reminder> reminders = reminderRepository.findByReminderDateBefore(new Date());
@@ -57,7 +40,7 @@ public class ReminderServiceImpl implements ReminderService {
             recipients.forEach(recipient -> {
                 mailService.sendEmailTemplate(template, properties, subject, Collections.singletonList(recipient));
                 logger.info("Send reminder {} to {}", reminder.getType(), recipient);
-                
+
                 Optional<User> userOptional = userRepository.findByEmail(recipient);
                 userOptional.ifPresent(user -> receiverIds.add(user.getId()));
             });
@@ -65,20 +48,15 @@ public class ReminderServiceImpl implements ReminderService {
             String title = groupOptional.isPresent() ? groupOptional.get().getName() : "MentorUS";
             String body = "";
             if (reminder.getType() == ReminderType.MEETING) {
-                body =
-                        "Lịch hẹn "
-                                + reminder.getName()
-                                + " sẽ diễn ra lúc "
-                                + reminder.getProperties().get("dueDate");
+                body = String.format("Lịch hẹn %s sẽ diễn ra lúc %s", reminder.getName(), reminder.getProperties().get("dueDate"));
             } else if (reminder.getType() == ReminderType.TASK) {
-                body =
-                        "Công việc "
-                                + reminder.getName()
-                                + " sẽ tới hạn lúc "
-                                + reminder.getProperties().get("dueDate");
+                body = String.format("Công việc %s sẽ tới hạn lúc %s", reminder.getName(), reminder.getProperties().get("dueDate"));
             }
             firebaseMessagingManager.sendGroupNotification(receiverIds, title, body);
         }
-        reminderRepository.deleteAll(reminders);
+
+        if (!reminders.isEmpty()) {
+            reminderRepository.deleteAll(reminders);
+        }
     }
 }
